Refactor imports to consolidate NostrEventProvider references
