+++ hooks/useAlbumComments.ts
@@ -1,6 +1,6 @@
 import { fetchContentCommentEvents, getAlbumTracks } from "@/utils";
 import { useQuery } from "@tanstack/react-query";
-import { nostrQueryKeys, useNostrEvents } from "@/providers/NostrEventProvider";
+import { nostrQueryKeys, useNostrEvents } from "@/providers";
 
 // this returns a list of event IDs for event kinds 1, 1985, and 9735
 export const useAlbumComments = (albumId: string, limit?: number) => {
